FossilOrigin-Name: d0305b305ad0a15a1aa38dbd1830bc9b7654eb69

+++ src/sha1.c
@@ -3,6 +3,7 @@
 ** contained in RFC-3174.
 */
 #include <stdint.h>
+#include <sys/types.h>
 #include "config.h"
 #include "sha1.h"
 
@@ -12,7 +13,6 @@
  *    name              meaning
  *  uint32_t         unsigned 32 bit integer
  *  uint8_t          unsigned 8 bit integer (i.e., unsigned char)
- *  int_least16_t    integer of >= 16 bits
  *
  */
 #define SHA1HashSize 20
@@ -31,8 +31,7 @@ struct SHA1Context {
     uint32_t Length_Low;            /* Message length in bits      */
     uint32_t Length_High;           /* Message length in bits      */
 
-                               /* Index into message block array   */
-    int_least16_t Message_Block_Index;
+    int Message_Block_Index;   /* Index into message block array   */
     uint8_t Message_Block[64];      /* 512-bit message blocks      */
 
     int Computed;               /* Is the digest computed?         */
